Move Environment to common package
COSDK-500
