+++ core/src/main/java/com/adyen/checkout/core/analytics/internal/AnalyticsManagerFactory.kt
@@ -10,7 +10,6 @@ package com.adyen.checkout.core.analytics.internal
 
 import android.app.Application
 import androidx.annotation.RestrictTo
-import com.adyen.checkout.core.Environment
 import com.adyen.checkout.core.analytics.internal.data.DefaultAnalyticsRepository
 import com.adyen.checkout.core.analytics.internal.data.local.ErrorAnalyticsLocalDataStore
 import com.adyen.checkout.core.analytics.internal.data.local.InfoAnalyticsLocalDataStore
@@ -19,6 +18,7 @@ import com.adyen.checkout.core.analytics.internal.data.remote.AnalyticsTrackRequ
 import com.adyen.checkout.core.analytics.internal.data.remote.DefaultAnalyticsRemoteDataStore
 import com.adyen.checkout.core.analytics.internal.data.remote.DefaultAnalyticsSetupProvider
 import com.adyen.checkout.core.analytics.internal.data.remote.api.AnalyticsService
+import com.adyen.checkout.core.common.Environment
 import com.adyen.checkout.core.common.internal.api.HttpClientFactory
 import com.adyen.checkout.core.components.data.model.Amount
 import com.adyen.checkout.core.internal.ui.model.ComponentParams
