+++ core/src/main/java/com/adyen/checkout/core/internal/Configuration.kt
@@ -10,8 +10,8 @@ package com.adyen.checkout.core.internal
 
 import android.os.Parcelable
 import androidx.annotation.RestrictTo
-import com.adyen.checkout.core.Environment
 import com.adyen.checkout.core.analytics.AnalyticsConfiguration
+import com.adyen.checkout.core.common.Environment
 import com.adyen.checkout.core.components.data.model.Amount
 import java.util.Locale
