+++ core/src/main/java/com/adyen/checkout/core/internal/ui/model/ComponentParams.kt
@@ -9,8 +9,8 @@
 package com.adyen.checkout.core.internal.ui.model
 
 import androidx.annotation.RestrictTo
-import com.adyen.checkout.core.Environment
 import com.adyen.checkout.core.analytics.internal.AnalyticsParams
+import com.adyen.checkout.core.common.Environment
 import com.adyen.checkout.core.components.data.model.Amount
 import java.util.Locale
