+++ core/src/main/java/com/adyen/checkout/core/internal/ui/model/SessionParams.kt
@@ -9,7 +9,7 @@
 package com.adyen.checkout.core.internal.ui.model
 
 import androidx.annotation.RestrictTo
-import com.adyen.checkout.core.Environment
+import com.adyen.checkout.core.common.Environment
 import com.adyen.checkout.core.components.data.model.Amount
 import java.util.Locale
