+++ core/src/main/java/com/adyen/checkout/core/sessions/CheckoutSession.kt
@@ -9,7 +9,7 @@
 package com.adyen.checkout.core.sessions
 
 import androidx.annotation.RestrictTo
-import com.adyen.checkout.core.Environment
+import com.adyen.checkout.core.common.Environment
 import com.adyen.checkout.core.components.CheckoutConfiguration
 import com.adyen.checkout.core.components.data.model.PaymentMethod
 import com.adyen.checkout.core.sessions.internal.data.model.SessionSetupResponse
