+++ core/src/main/java/com/adyen/checkout/core/sessions/CheckoutSessionProvider.kt
@@ -8,7 +8,7 @@
 
 package com.adyen.checkout.core.sessions
 
-import com.adyen.checkout.core.Environment
+import com.adyen.checkout.core.common.Environment
 import com.adyen.checkout.core.internal.Configuration
 
 internal object CheckoutSessionProvider {
