+++ example-app/src/main/java/com/adyen/checkout/example/ui/v6/V6ViewModel.kt
@@ -9,7 +9,7 @@
 package com.adyen.checkout.example.ui.v6
 
 import androidx.lifecycle.ViewModel
-import com.adyen.checkout.core.Environment
+import com.adyen.checkout.core.common.Environment
 import com.adyen.checkout.core.components.AdyenCheckout
 import com.adyen.checkout.core.components.CheckoutCallback
 import com.adyen.checkout.core.components.CheckoutConfiguration
